fix init pcb error, multicore task runnable

+++ UCAS_OS/Project3-InteractiveProcMangaer/include/os/lock.h
@@ -69,11 +69,11 @@ int spin_lock_try_acquire(spin_lock_t *lock);
 void spin_lock_acquire(spin_lock_t *lock);
 void spin_lock_release(spin_lock_t *lock);
 
-long k_mutex_lock_op(long *key,int op);
+long k_mutex_lock_op(int *key,int op);
 long k_mutex_lock_init(int *key, int operator);
-long k_mutex_lock_acquire(long key, int operator);
-long k_mutex_lock_release(long key, int operator);
-long k_mutex_lock_destroy(long *key, int operator);
+long k_mutex_lock_acquire(int key, int operator);
+long k_mutex_lock_release(int key, int operator);
+long k_mutex_lock_destroy(int *key, int operator);
 long k_mutex_lock_trylock(int *key, int operator);
 
 #endif
